requirements.txt
python-telegram-bot==20.7
yt-dlp==2025.5.22
python-dotenv==1.1.0
